rename C# indexer to anilibria, delete yaml, fix bugs
